% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.table.R
\name{c.table}
\alias{c.table}
\title{Title c.table (contingency table)}
\usage{
\method{c}{table}(
  df,
  x,
  y,
  col_percent = TRUE,
  row_percent = FALSE,
  show_totals = TRUE,
  chi = TRUE,
  decimals = 2,
  debug = FALSE
)
}
\arguments{
\item{df}{dataframe to be used}

\item{x}{variable that will be on the columns}

\item{y}{variable that will be on the rows}

\item{col_percent}{T/F show percentage by columns}

\item{row_percent}{T/F show percentage by rows}

\item{show_totals}{T/F show totals}

\item{chi}{T/F perform proportions comparison (chi.square or fisher, as needed)}

\item{decimals}{number of decimal digits}
}
\value{

}
\description{
Title c.table (contingency table)
}
