% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_means.R
\name{c_means}
\alias{c_means}
\title{Title}
\usage{
c_means(
  df,
  x,
  group,
  decimals = 2,
  alternative = "two.sided",
  debug = FALSE,
  show.test.method = TRUE,
  conf.level = 0.95,
  show.warnings = TRUE,
  is.normal = TRUE
)
}
\arguments{
\item{is.normal}{}
}
\value{

}
\description{
Title
}
