% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{correlation}
\usage{
correlation(
  data,
  ...,
  decimals = 2,
  show.errors = TRUE,
  show.warnings = TRUE,
  DEBUG = FALSE,
  p.sig = 0.05,
  spearman.cutoff = c(0.25, 0.5, 0.7),
  spearman.labels = c("weak", "moderate", "strong"),
  pearson.cutoff = c(0.25, 0.5, 0.7),
  pearson.labels = c("weak", "moderate", "strong"),
  kendall.cutoff = c(0.25, 0.5, 0.7),
  kendall.labels = c("weak", "moderate", "strong"),
  inv.string = "(inv)",
  method = c("auto", "pearson", "spearman", "kendall"),
  lang = c("en", "es")
)
}
\arguments{
\item{lang}{}
}
\value{

}
\description{
correlation
}
