% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{.do.means}
\alias{.do.means}
\title{Title do.means}
\usage{
.do.means(
  variable,
  show.errors = FALSE,
  DEBUG = TRUE,
  decimales = 2,
  n = TRUE,
  missing = TRUE,
  min = TRUE,
  max = TRUE,
  mean = TRUE,
  sd = TRUE,
  median = TRUE,
  IQR = TRUE,
  norm.test = TRUE
)
}
\arguments{
\item{norm.test}{}
}
\value{

}
\description{
This is the function that does the calculation per se
}
