% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{.new_correlation}
\alias{.new_correlation}
\title{Title}
\usage{
.new_correlation(
  original.data = data.frame(),
  main = data.frame(),
  corr.matrix = data.frame(),
  p.matrix = data.frame(),
  method.matrix = data.frame(),
  method = c("auto", "sw", "lillie", "ks"),
  theory.text = character(),
  interpretation.matrix = data.frame()
)
}
\arguments{
\item{interpretation.matrix}{}
}
\value{

}
\description{
Title
}
