% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.col.R
\name{find.col}
\alias{find.col}
\title{Title}
\usage{
find.col(data, string = "", show.col.number = FALSE)
}
\arguments{
\item{data}{data.frame used}

\item{string}{quoted string to search for (doesn't need wildcards) "dat" will look for "data" as well as "date"}

\item{show.col.number}{sometimes it's usefull to know the number of the columns found, this is where you ask for it}
}
\value{
list of names that contains the string with or without the column number besides it
}
\description{
Title
}
