% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{is.normal}
\alias{is.normal}
\title{Check for normality
is.normal() gives a TRUE/FALSE value based on normality test of the variable.}
\usage{
is.normal(variable, show.warnings = FALSE, decimals = 2, DEBUG = FALSE)
}
\arguments{
\item{variable}{vector of values (ideally numbers)}

\item{show.warnings}{Default FALSE. Whether or not you want to get possible warnings}
}
\value{
logical value (TRUE/FALSE)
}
\description{
Check for normality
is.normal() gives a TRUE/FALSE value based on normality test of the variable.
}
\examples{

is.normal(c(1,76,3,2,1,23,5,64,234,12,3,43,2324,54,45,223,341,213,2,4323,24,3,54,54,56,76,7,55,3,2323))

}
