% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{media}
\alias{media}
\title{Title media}
\usage{
media(
  df,
  ...,
  digits = 2,
  n = TRUE,
  missing = TRUE,
  min = TRUE,
  max = TRUE,
  mean = TRUE,
  sd = TRUE,
  q1 = TRUE,
  median = TRUE,
  q3 = TRUE,
  iq_range = TRUE
)
}
\arguments{
\item{...}{any number of variables either as vector or as column names of data.frame see Examples}

\item{n}{show n (without missing values)}

\item{missing}{show number of missing values}

\item{min}{show min value (or NA)}

\item{max}{show max value (or NA)}

\item{mean}{show mean value}

\item{sd}{show standard deviation}

\item{median}{show median}

\item{data}{data.frame with variables or vector}

\item{by}{column of data.frame to sort the results by (either as VARIABLE or as "VARIABLE")}

\item{decimales}{number of decimals}

\item{show.warnings}{show warnings or not}

\item{show.errors}{show errors or not}

\item{IQR}{show interquantile range}

\item{norm.test}{show normality test}

\item{result.as.data.frame}{return result as a data.frame or as a list}
}
\value{
data.frame (or list) with result/s
}
\description{
Title media
}
\examples{

data_ <- data.frame(AGE=sample(x = 65:100, size=30, replace = TRUE ),
                    HEIGHT=sample(x = 120:205, size=30, replace = TRUE ),
                    SEX=sample(x = c("Male", "Female"), prob = c(.5,.5), size = 30, replace = TRUE),
                    BLOND=sample(x = c("Yes", "No"), prob = c(.2,.8), size = 30, replace = TRUE)
)
data_ <- rbind(data_, list(NA,NA,NA,NA))
data_ <- rbind(data_, list(NA,NA,"Male",NA))
data_ <- rbind(data_, list(NA,NA,NA,"No"))

media(df$AGE)

media(df$AGE, df$HEIGHT)

media(df,AGE)
media(df,AGE,HEIGHT)

media(df,"AGE")
media(df,"AGE","HEIGHT")

media(df,AGE, by=SEX)



}
