% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{norm.test}
\alias{norm.test}
\title{Normality Test}
\usage{
norm.test(
  variable,
  to.numeric = TRUE,
  decimals = 2,
  method = "auto",
  lang = "en",
  show.interpretation = FALSE,
  show.theory = FALSE,
  show.warnings = TRUE
)
}
\arguments{
\item{variable}{vector (numeric if possible) of observations to for wich normality will be assessed}

\item{to.numeric}{if the vector is not numeric: do you want to try an automatic conversion?}

\item{decimals}{number of decimals for the p value (it will determine the threshold for "<x.xxx1" p value too)}

\item{method}{string with the possible methods: "sw", "lille", "ks" (default: "auto")}

\item{lang}{either "en" for english or "es" for spanish}

\item{show.interpretation}{a logical value indicating if you want an interpretation of normality and the specific result you have}

\item{show.theory}{a logical value indicating if you want a background explanation of the test}

\item{show.warnings}{a logical value indicating wether you want warnings to be shown or not}
}
\value{
It returns list of elements:
 * var                       name of the variable tested
 * method                    method used to determine normality
 * p.value                   exact p value (rounded as per decimals argument)
 * p.value.str               STRING value with p value. If p value < 10^-(decimals + 1) then returns "<0.(decimals)1". If decimals = 2 then <0.001
 * normal                    logical value indicating if the variable follows a normal distribution or not
 * theory                    TEXT theory about normality test
 * interpret                 TEXT interpreting the results
 * config:
    + show.theory            whether or not a theory explanation was solicited
    + show.interpretation    whether or not an interpretation was solicited
    + lang                   languaje selected for the output
}
\description{
This function assess if a variable follows the normal distribution. This is done using Shapiro-Wilk's test whenever possible and Lilliefor's correction for Kolmogorov-Smirnoff if Shapiro-Wilk's is not possible.
}
