% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecha.spss.R
\name{pss2datetime}
\alias{pss2datetime}
\title{Title pss2datetime}
\usage{
pss2datetime(x)
}
\arguments{
\item{x}{Fecha en formato SPSS con HORA}
}
\value{
Fecha en formato R
}
\description{
Para obtener la fecha con horas minutos y segundos
}
