% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemann_sum.R
\name{riemann_sum}
\alias{riemann_sum}
\title{riemann.sum}
\usage{
riemann_sum(
  df,
  ...,
  group_by_col = NULL,
  inc.value = NULL,
  min.value = NULL,
  max.value = NULL,
  debug = FALSE
)
}
\arguments{
\item{...}{variables to be studied (no quotes)}

\item{group_by_col}{name of the variable for the grouping (if there is one)}

\item{inc.value}{split the studied variable in increments of x}

\item{min.value}{min value of x for witch the riemann.sum will estimate the area under the curve}

\item{max.value}{max value of x for witch the riemann.sum will estimate the area under the curve}

\item{debug}{debuggin options on or off}
}
\value{
dataframe with values
}
\description{
This function calculates the area under the curve for a density function.
}
