% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecha.spss.R
\name{spss.date.time}
\alias{spss.date.time}
\title{Title spss.date.time}
\usage{
spss.date.time(x)
}
\arguments{
\item{x}{Fecha en formato SPSS con HORA}
}
\value{
Fecha en formato R
}
\description{
Para obtener la fecha con horas minutos y segundos
}
