% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{means}
\alias{means}
\title{=============================================================
 TO BE REMOVED: this is here just for compatibility
=============================================================}
\usage{
means(
  df,
  ...,
  group_by_col = NULL,
  decimales = 2,
  show_warnings = TRUE,
  show_errors = FALSE,
  n = TRUE,
  missing = TRUE,
  min = TRUE,
  max = TRUE,
  mean = TRUE,
  sd = TRUE,
  median = TRUE,
  range = TRUE,
  norm.test = TRUE,
  col_names = c("var", "groups", "n.valid", "n.missing", "min", "max", "mean", "sd",
    "median", "range", "shapiro-wilk")
)
}
\arguments{
\item{df}{data}

\item{...}{any number of numerical variables}

\item{group_by_col}{variable to group the others by}

\item{decimales}{number of decimal to show}

\item{show_warnings}{show warnings}

\item{show_errors}{show errors}

\item{n}{show n (without missing values)}

\item{missing}{show number of missing values}

\item{min}{show min value (or NA)}

\item{max}{show max value (or NA)}

\item{mean}{show mean value}

\item{sd}{show standard deviation}

\item{median}{show median}

\item{range}{show range}

\item{norm.test}{perform Shapiro-Wilks test and show p.value}

\item{col_names}{column names for all columns (must set all or none)}
}
\value{
data.frame with the corresponding stats
}
\description{
Title means
}
\details{
My own spin on the mean function, working as I like it
}
\examples{

data_ <- data.frame(AGE=sample(x = 65:100, size=30, replace = TRUE ),
                    HEIGHT=sample(x = 120:205, size=30, replace = TRUE ),
                    SEX=sample(x = c("Male", "Female"), prob = c(.5,.5), size = 30, replace = TRUE),
                    BLOND=sample(x = c("Yes", "No"), prob = c(.2,.8), size = 30, replace = TRUE)
)
data_ <- rbind(data_, list(NA,NA,NA,NA))
data_ <- rbind(data_, list(NA,NA,"Male",NA))
data_ <- rbind(data_, list(NA,NA,NA,"No"))

means(data_, AGE)

means(data_, AGE,HEIGHT)

mean(data_, AGE, group_by_col=SEX)

}
